ParticleProperties: Read list of QuantumNumbers from XML file
